Performing changes to controller method, deleting try and catch block because of added exception handler

+++ centersystem/src/main/java/clinic/centersystem/service/ClinicCenterAdminServiceCont.java
@@ -75,6 +75,10 @@ public class ClinicCenterAdminServiceCont {
     public String approveRegistrationRequest(Long id) {
         RegistrationRequirement req = registrationRequirementService.findById(id);
         req.setPassword(passwordEncoder.encode(req.getPassword()));
+        if (this.userService.existsByEmail(req.getEmail())) {
+            System.out.println("exists");
+            throw new UserExistsException();
+        }
         Patient patient = this.patientService.save(req);
         this.registrationRequirementService.deleteById(id);
